refactor the auth process
