+++ st2common/st2common/util/pecan.py
@@ -1,0 +1,27 @@
+# Licensed to the StackStorm, Inc ('StackStorm') under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+import pecan
+from six.moves import http_client
+
+__all__ = [
+    'abort_request',
+]    
+
+
+def abort_request(status_code=http_client.UNAUTHORIZED,
+                  message='Invalid or missing credentials'):
+    pecan.abort(status_code, message)
